third_party/blink/web_tests/wpt_internal/attribution-reporting/register-attribution-source.https.html

<!doctype html>
<title>window.attributionReporting.registerAttributionSource</title>
<link rel="help" href="https://github.com/WICG/conversion-measurement-api/blob/main/event_attribution_reporting.md#registering-attribution-sources-with-javascript">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(t => {
  return window.attributionReporting.registerAttributionSource({
    attributionSourceEventId: '1',
    attributionDestination: 'https://a.example',
    attributionReportTo: 'https://b.example',
  });
}, 'valid params');

promise_test(t => {
 return promise_rejects_js(t, TypeError,
    window.attributionReporting.registerAttributionSource({
      attributionSourceEventId: '!',
      attributionDestination: 'https://a.example',
    }));
}, 'invalid attributionSourceEventId');

promise_test(t => {
 return promise_rejects_js(t, TypeError,
    window.attributionReporting.registerAttributionSource({
      attributionSourceEventId: '1',
      attributionDestination: 'http://a.example',
    }));
}, 'insecure attributionDestination');

promise_test(t => {
 return promise_rejects_js(t, TypeError,
    window.attributionReporting.registerAttributionSource({
      attributionSourceEventId: '1',
      attributionDestination: 'https://a.example',
      attributionReportTo: 'http://b.example',
    }));
}, 'insecure attributionReportTo');

</script>
